==> before_everything_there_were_notebooks/04_DoPredictionForBoss.ipynb <==
<hr/>
<p style="font-size: 1.3em; color: red;">START: Inference</p>

# Inference and scoring

👨‍💻: This is our prediction and scoring notebook. Simply load your model and we'll do the rest.

! pip freeze -l

👨‍💻: I should probably save the above requirements at some point...

import pandas as pd

df_to_predict = pd.read_csv('data/csv/bank_marketing_2010-08-01_to_2010-09-30.csv', sep=';')

# Magistral mistake: Let's pretend we don't have y
df_to_predict = df_to_predict.drop(columns=['y','duration'])

Let's load our models

import xgboost as xgb
import pickle

clf = xgb.Booster(model_file='best_xgb_clf.model')

with open('best_xgb_data_processing.pkl','rb') as f:
    processing = pickle.load(f)

out = clf.predict(xgb.DMatrix(processing.transform(df_to_predict)))

<p style="font-size: 1.3em; color: red;">End of section: Inference STOP HERE AND GO BACK TO INSTRUCTIONS</p>
<hr/>

<hr/>
<p style="font-size: 1.3em; color: red;">START: Scoring</p>

👨‍💻: This is our scoring!

# TODO: Complete me

<p style="font-size: 1.3em; color: red;">End of section: Scoring STOP HERE AND GO BACK TO INSTRUCTIONS</p>
<hr/>
